+++ contracts/handlers/furucombo/funds/IFunds.sol
@@ -7,10 +7,10 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 interface IFunds{
     function denomination() external view returns (address);
     function shareToken() external view returns (address);
-    function calculateShare(uint256) external view returns (uint256);
-    function calculateBalance(uint256) external view returns (uint256);
+    function calculateShare(uint256 balance_) external view returns (uint256 share);
+    function calculateBalance(uint256 share_) external view returns (uint256 balance);
     function vault() external view returns (address);
 
-    function purchase(uint256) external returns (uint256);
-    function redeem(uint256, bool) external returns (uint256);
+    function purchase(uint256 balance_) external returns (uint256 share);
+    function redeem(uint256 share_, bool acceptPending_) external returns (uint256 balance);
 }